feat(editor): wire drag UX — bidirectional, sticky chain, disconnect
* Drag from either output or input — edge stored in canonical output→input order regardless of which side started
* Hold-and-release model is default (drop sticky default that confused users)
* Shift on output pin Press → sticky chain: release commits and source stays for fan-out to many inputs; click-empty / right-click ends chain
* Right-click on any pin → disconnect all edges touching it
* HitTester / hover walk: non-clipping nodes no longer reject overflow children, so pin handles straddling the card border are clickable across their full visual area (caller verifies own bounds only before invoking own handlers)
* Move events now dispatched to handlers via root.hitTest so a sticky wire can follow the cursor with no button held
* WireLayer renders rubber-band oriented output→input regardless of grab side

+++ src/main/kotlin/dev/nitka/nodewire/ui/core/NwComposeScreen.kt
@@ -93,7 +93,7 @@ abstract class NwComposeScreen(title: Component) : Screen(title) {
     }
 
     override fun mouseDragged(x: Double, y: Double, btn: Int, dx: Double, dy: Double): Boolean {
-        if (owner.dispatchPointer(PointerEvent.Drag(x.toInt(), y.toInt(), btn, dx.toInt(), dy.toInt()))) return true
+        if (owner.dispatchPointer(PointerEvent.Drag(x.toInt(), y.toInt(), btn, dx.toFloat(), dy.toFloat()))) return true
         return super.mouseDragged(x, y, btn, dx, dy)
     }
